+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - added keyboard shortcuts
 - fixed task selection on touch devices
 - added navigation for jumping between user messages
+- correctly showing usage data on tool results when loading existing tasks
 
 ## [0.45.1]